feat(web/comments): Slice 2 — comment moderation with optimistic publish
Wires moderate() into createComment with the design from
dev-docs/policy-moderator-plan.md §3 + §7.2:

  - Hard block on 'illegal'. The comment row is never inserted;
    a policy_decisions row goes in with target_id=NULL, the user
    gets a notification, and checkBanCandidate files a ban_candidate
    flag against the parent submission so /admin/queue surfaces it
    under SLA. Returns reason='illegal' on the new CommentResult
    union so REST clients map to 422.

  - Optimistic publish on non-illegal rejects (spam/abuse/doxxing).
    The comment is inserted immediately to keep threads alive, then
    Vercel after() schedules a confirmation pass. The confirmation
    re-runs moderate() against the freshly-inserted row; if it still
    rejects, comments/confirm.ts retracts (state='rejected'),
    writes the pass=2 policy_decisions row, the moderation_log row,
    and the notification. If the confirmation passes, the comment
    stands and the discrepancy is captured in the two-row audit
    trail as a calibration signal.

  - Pass verdicts: existing flow, plus a pass-1 policy_decisions
    row written outside the FOR-SHARE transaction.

Critical: moderate() runs BEFORE the transaction. A 1500ms model
call inside the transaction would hold parent locks and serialize
comment writes on hot threads. The transaction stays narrow — just
target/parent FOR-SHARE locks, the insert, and the reply
notification.

Ladder integration:
  - checkLadderRateLimit at the top: rung 3 daily cap drops after
    accumulating rejects. Returns the new 'rate' reason on
    CommentResult so callers map to 429.
  - checkBanCandidate after each terminal reject (illegal block,
    pass-2 retract): rung 4 ban-candidate flag. The submission and
    comment routes now both return 429 on rate-limited rejects with
    a UTC-midnight reset.

Splits the prior 346-LOC lib/comments.ts into a directory matching
the lib/submissions/ pattern (schema, state, create, edit, delete,
confirm, index). Existing four importers (REST, MCP, web action)
resolve through the new barrel unchanged. comments/create.ts is
269 LOC — the heaviest file, but well under the 350 ceiling.

No DB migration. Schema additions all came in 0018.
