+++ web/src/lib/comments/edit.ts
@@ -1,0 +1,100 @@
+/**
+ * Author-only edit. Window policy mirrors updateSubmissionAsAuthor:
+ *
+ *   Authorization (who can edit, when):
+ *     - Humans (role=user, is_agent=false): only within 5-min window.
+ *     - Bots / system / staff: any time.
+ *
+ *   Visibility (does the edit show as "edited"):
+ *     - Within-window edits are SILENT (no updated_at bump) for
+ *       everyone, including bots. The window is the period in which
+ *       no reader could have seen the original; a correction inside
+ *       it is a typo fix.
+ *     - Out-of-window edits bump updated_at and render the badge.
+ *
+ * Soft-deleted comments are not editable — the row is rendered as
+ * a "[deleted]" tombstone and editing it would surface the original
+ * body under the deleted-author label.
+ */
+
+import { revalidatePath } from "next/cache";
+import { and, eq, sql } from "drizzle-orm";
+
+import { db } from "@/db/client";
+import { comments, users } from "@/db/schema";
+import type { UpdateCommentInput, UpdateCommentResult } from "./schema";
+
+const EDIT_WINDOW_MS = 5 * 60 * 1000;
+
+export async function updateCommentAsAuthor(
+  authorId: string,
+  commentId: string,
+  input: UpdateCommentInput,
+): Promise<UpdateCommentResult> {
+  const [actor] = await db
+    .select({ role: users.role, isAgent: users.isAgent })
+    .from(users)
+    .where(eq(users.id, authorId))
+    .limit(1);
+  if (!actor) return { ok: false, reason: "not_found" };
+
+  const [existing] = await db
+    .select({
+      authorId: comments.authorId,
+      submissionId: comments.submissionId,
+      createdAt: comments.createdAt,
+      body: comments.body,
+      deletedAt: comments.deletedAt,
+    })
+    .from(comments)
+    .where(eq(comments.id, commentId))
+    .limit(1);
+  if (!existing) return { ok: false, reason: "not_found" };
+  // Soft-deleted comments stay readable as tombstones; editing one
+  // would surface the original body as if undeleted. Treat as gone.
+  if (existing.deletedAt) return { ok: false, reason: "not_found" };
+  if (existing.authorId !== authorId) return { ok: false, reason: "forbidden" };
+
+  const ageMs = Date.now() - existing.createdAt.getTime();
+  const withinWindow = ageMs <= EDIT_WINDOW_MS;
+  const bypassesWindow =
+    actor.isAgent || actor.role === "system" || actor.role === "staff";
+  if (!withinWindow && !bypassesWindow) {
+    return { ok: false, reason: "expired" };
+  }
+  if (input.body === existing.body) return { ok: false, reason: "noop" };
+
+  // Visibility = function of time, not role. See header.
+  const silent = withinWindow;
+  const bumpedAt = silent ? null : new Date();
+
+  // Atomic guard: re-check authorship + not-deleted in the WHERE so
+  // a concurrent delete or role-flip can't slip a write through. Omit
+  // updatedAt from .set() when silent so any prior post-window bump
+  // is preserved (defensive against future re-edit paths).
+  const updates: Partial<{ body: string; updatedAt: Date }> = {
+    body: input.body,
+  };
+  if (!silent) updates.updatedAt = bumpedAt as Date;
+
+  const updated = await db
+    .update(comments)
+    .set(updates)
+    .where(
+      and(
+        eq(comments.id, commentId),
+        eq(comments.authorId, authorId),
+        sql`${comments.deletedAt} IS NULL`,
+      ),
+    )
+    .returning({ id: comments.id });
+  if (updated.length === 0) return { ok: false, reason: "not_found" };
+
+  revalidatePath(`/post/${existing.submissionId}`);
+  return {
+    ok: true,
+    silent,
+    submissionId: existing.submissionId,
+    updatedAt: bumpedAt,
+  };
+}
